chore: redistribute shared wallet connect files
