+++ packages/apps/human-app/frontend/src/shared/hooks/use-wallet-connect.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { WalletConnectContext } from '@/shared/contexts/wallet-connect';
+import { WalletConnectContext } from '../contexts/wallet-connect';
 
 export const useWalletConnect = () => {
   const context = useContext(WalletConnectContext);
