+++ packages/apps/human-app/frontend/src/shared/interfaces/wallet-connect-account-state.interface.ts
@@ -1,0 +1,13 @@
+export interface ConnectedAccount {
+  isConnected: true;
+  chainId: number;
+  address: string;
+  signMessage: (message: string) => Promise<string | undefined>;
+}
+
+export interface DisconnectedAccount {
+  isConnected: false;
+  chainId?: never;
+  address?: never;
+  signMessage?: (message: string) => Promise<string | undefined>;
+}
